chore(launchpad-g/erc20): rename create token/airdrop/sale screens files
